rv-runtime-generator/linker: Align end of NAPOT regions after emitting the last section
Currently, rv-runtime-generator aligns the last section in a NAPOT
region to the size of that region. This makes the section blow up in
size even if it doesn't really require all that space. Instead of
putting this space in the last section, this change adds a
post_section_alignment_in_bytes as an optional member in SectionInfo
that is set only for the last section in a NAPOT region and generates
align statement after the last section is emitted to align the
location counter to end of napot region.